chore: bump extensions ref in boot assets image
Get the boot assets image (the image containing the agent extension, used in development of the Omni bare-metal infra provider) to be built with the changes introduced in: https://github.com/siderolabs/extensions/pull/595.

We need this for the development/testing processes of the bare-metal infra provider to introduce the "TLS connectivity to the agents" functionality.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>